+++ examples/empty/src/Controller.hpp
@@ -1,4 +1,4 @@
 #pragma once
 # include <Siv3D.hpp>
 
-Array<int32> getWherePush();
+Array<int8> getWherePush();
